webpage/inchcape/requirements/base.txt
Django==1.10
defusedxml==0.4.1
django-allauth==0.26.1
django-model-utils==2.5.2
oauthlib==1.1.2
psycopg2==2.6.2
python3-openid==3.0.10
requests==2.11.0
requests-oauthlib==0.6.2
